Time: 64 ms (41.53%), Space: 27.4 MB (92.24%) - LeetHub

+++ 704-binary-search/704-binary-search.cpp
@@ -6,8 +6,8 @@ public:
         
         
         
-        int left = 0;
-        int right = nums.size() - 1;
+        int16_t left = 0;
+        int16_t right = nums.size() - 1;
         
        while(left <= right){
